Fix Phase 1 build: re-encode TracableRectangle as UTF-8; review fixes
- TracableRectangle.cs was rewritten but inherited the original file's
  UTF-16 encoding, leaving embedded NUL bytes that broke the compile
  (CS1056). Re-saved as UTF-8.
- GpuPhase1Tests.cs: `int / (uint)` promotes to long and can't implicitly
  assign to uint (CS0266). Cast the int/int division instead (two spots).
- PathTrace.hlsl / Phase1Renderer: add a per-pixel LastHit buffer and
  restart the running mean when a pixel flips hit<->miss, matching the
  CPU TraceCore 'hit != LastHit' reset so silhouette-edge pixels converge
  the same way under sub-pixel jitter.

+++ RayTracer/RayTracer.Tests/GpuPhase1Tests.cs
@@ -217,7 +217,7 @@ public class GpuPhase1Tests
                 foreach (var (pixelHash, sampleIdx) in samples)
                 {
                     uint heroIdx = Phase1Reference.HeroIndex(pixelHash, sampleIdx, DeterministicCount);
-                    uint stride = DeterministicCount / (uint)Phase1Reference.CompanionCount;
+                    uint stride = (uint)(DeterministicCount / Phase1Reference.CompanionCount);
 
                     // Hero hit gives us the exact hit point the shader would reconstruct UV from.
                     var heroRay = new Ray
@@ -235,7 +235,7 @@ public class GpuPhase1Tests
                     Vector3 baseXyzTruth = Vector3.Zero;
                     for (uint k = 0; k < Phase1Reference.CompanionCount; k++)
                     {
-                        uint idx = (heroIdx + k * stride) % DeterministicCount;
+                        uint idx = (heroIdx + k * stride) % (uint)DeterministicCount;
                         int wlK = res.DeterWavelengths[idx];
                         var ray = new Ray { Origin = origin, Direction = dir, Wavelength = wlK, Intensity = 1f };
                         var hit = tracable.Intersect(ray);
